+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/rest/AbstractReadDtoController.java
@@ -33,6 +33,7 @@ import eu.bcvsolutions.idm.core.api.config.swagger.SwaggerConfig;
 import eu.bcvsolutions.idm.core.api.dto.BaseDto;
 import eu.bcvsolutions.idm.core.api.dto.filter.BaseFilter;
 import eu.bcvsolutions.idm.core.api.dto.filter.DataFilter;
+import eu.bcvsolutions.idm.core.api.dto.filter.PermissionContext;
 import eu.bcvsolutions.idm.core.api.exception.EntityNotFoundException;
 import eu.bcvsolutions.idm.core.api.exception.ForbiddenEntityException;
 //import eu.bcvsolutions.idm.core.api.exception.ResultCodeException;
@@ -41,6 +42,7 @@ import eu.bcvsolutions.idm.core.api.service.ReadDtoService;
 import eu.bcvsolutions.idm.core.api.utils.FilterConverter;
 import eu.bcvsolutions.idm.core.security.api.domain.BasePermission;
 import eu.bcvsolutions.idm.core.security.api.domain.IdmBasePermission;
+import eu.bcvsolutions.idm.core.security.api.utils.PermissionUtils;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
@@ -159,7 +161,7 @@ public abstract class AbstractReadDtoController<DTO extends BaseDto, F extends B
 	public Resources<?> find(
 			@RequestParam(required = false) MultiValueMap<String, Object> parameters,
 			@PageableDefault Pageable pageable) {
-		return toResources(find(toFilter(parameters), pageable, IdmBasePermission.READ), getDtoClass());
+		return toResources(find(toFilter(parameters), pageable, evaluatePermission(parameters, IdmBasePermission.READ)), getDtoClass());
 	}
 	
 	/**
@@ -215,7 +217,7 @@ public abstract class AbstractReadDtoController<DTO extends BaseDto, F extends B
 	public Resources<?> autocomplete(
 			@RequestParam(required = false) MultiValueMap<String, Object> parameters,
 			@PageableDefault Pageable pageable) {
-		return toResources(find(toFilter(parameters), pageable, IdmBasePermission.AUTOCOMPLETE), getDtoClass());
+		return toResources(find(toFilter(parameters), pageable, evaluatePermission(parameters, IdmBasePermission.AUTOCOMPLETE)), getDtoClass());
 	}
 	
 	/**
@@ -230,7 +232,7 @@ public abstract class AbstractReadDtoController<DTO extends BaseDto, F extends B
 			@Authorization(SwaggerConfig.AUTHENTICATION_CIDMST)
 			})
 	public long count(@RequestParam(required = false) MultiValueMap<String, Object> parameters) {
-		return count(toFilter(parameters), IdmBasePermission.COUNT);
+		return count(toFilter(parameters), evaluatePermission(parameters, IdmBasePermission.COUNT));
 	}
 
 	/**
@@ -358,5 +360,14 @@ public abstract class AbstractReadDtoController<DTO extends BaseDto, F extends B
 	 */
 	protected DTO checkAccess(DTO dto, BasePermission... permission) {
 		return getService().checkAccess(dto, permission);
+	}
+	
+	protected BasePermission evaluatePermission(MultiValueMap<String, Object> parameters, BasePermission originalPermission) {
+		// We need to use raw parameters => data filter (~PermissionContext instance) is not required now.
+		BasePermission permission = PermissionUtils.toPermission(
+				getParameterConverter().toString(parameters, PermissionContext.PARAMETER_EVALUATE_PERMISSION)
+		);
+		//
+		return permission == null ? originalPermission : permission;
 	}
 }