+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/security/api/utils/PermissionUtils.java
@@ -9,12 +9,12 @@ import java.util.stream.Collectors;
 
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.EnumUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.util.ObjectUtils;
 
 import com.google.common.collect.Lists;
 
 import eu.bcvsolutions.idm.core.security.api.domain.BasePermission;
-import eu.bcvsolutions.idm.core.security.api.domain.IdentityBasePermission;
 import eu.bcvsolutions.idm.core.security.api.domain.IdmBasePermission;
 
 /**
@@ -77,17 +77,15 @@ public abstract class PermissionUtils {
 	}
 
 	/**
-	 * Method resolve given list of permission constants (eq. 'IDENTITY_READ').
+	 * Method resolve given list of permission constants (e.g. 'IDENTITY_READ').
 	 * From the list will be parsed (by separator from BasePermission) permission. For IDENTITY_READ
 	 * will be result READ.
-	 * BasePermission enum will be get from these enums: {@link IdmBasePermission} or {@link IdentityBasePermission}.
-	 * If constant isn't found method throws error.
 	 *
 	 * BEWARE: if given list contains constant from different group result set will be united. 
 	 * For input list IDENTITY_READ, ROLE_UPDATE, ROLE_READ will be result: READ, UPDATE!
 	 *
-	 * @param permissions
-	 * @return BasePermission list 
+	 * @param authorities group (authorities) or base permissions.
+	 * @return BasePermission list of base permissions
 	 */
 	public static Collection<BasePermission> toPermissions(Collection<String> authorities) {
 		if (CollectionUtils.isEmpty(authorities)) {
@@ -96,27 +94,49 @@ public abstract class PermissionUtils {
 		Set<String> resolvedPermissions = new HashSet<>(authorities.size());
 		Set<BasePermission> result = new HashSet<>(authorities.size());
 		for (String authority : authorities) {
-			if (authority.contains(BasePermission.SEPARATOR)) {
-				String[] split = authority.split(BasePermission.SEPARATOR);
-				// permission is on last place
-				authority = split[split.length - 1];
-			}
-			if (resolvedPermissions.contains(authority)) {
+			BasePermission permission = toPermission(authority);
+			String permissionName = permission.getName();
+			if (resolvedPermissions.contains(permissionName)) {
 				continue;
-			}
-			final String rawPermission = authority;
-			// Base permission may be child from IdmBasePermission
-			BasePermission permission = EnumUtils.getEnum(IdmBasePermission.class, rawPermission);
-			// but can be registered dynamically in custom module => new BasePermission is created
-			if (permission == null) {
-				permission = (BasePermission) () -> rawPermission;
 			}
 			//
 			result.add(permission);
-			resolvedPermissions.add(permission.getName());
+			resolvedPermissions.add(permissionName);
 		}
 		return result;
 	}
+	
+	/**
+	 * Method resolve given permission constants (e.g. 'IDENTITY_READ').
+	 * Value will be parsed (by separator from BasePermission) permission. For 'IDENTITY_READ'
+	 * will be result 'READ'.
+	 * 
+	 * @param authority group (authority) or base permission.
+	 * @return base permission
+	 * @since 10.3.0
+	 */
+	public static BasePermission toPermission(String authority) {
+		if (StringUtils.isEmpty(authority)) {
+			return null;
+		}
+		//
+		if (authority.contains(BasePermission.SEPARATOR)) {
+			String[] split = authority.split(BasePermission.SEPARATOR);
+			// permission is on last place
+			authority = split[split.length - 1];
+		}
+		//
+		final String rawPermission = authority;
+		// Base permission may be child from IdmBasePermission
+		BasePermission permission = EnumUtils.getEnum(IdmBasePermission.class, rawPermission);
+		// but can be registered dynamically in custom module => new BasePermission is created
+		if (permission == null) {
+			permission = (BasePermission) () -> rawPermission;
+		}
+		//
+		return permission;
+	}
+	
 	/**
 	 * Converts set of {@link BasePermission} to set of permission names.
 	 *